feat: add color & Blinn-Phong example (#16)
* feat(lighting): init color example

* feat(lighting): add color example

* style: fmt

* test: add nalgebra test

* feat(lighting): init Blinn-Phong example

* fix: repair blinn-phong example & camera

* feat(lighting): add normal matrix

* fix(camera): repair camera rotation logic

* fix: clippy warnings

* feat(lighting): complete blinn-phong example

* style: fmt

* :sparkles: feat(lighting): add material example

+++ src/window_glfw.rs
@@ -50,7 +50,7 @@ impl GlfwWindow {
 
         let (width, height) = inner_win.get_framebuffer_size();
         unsafe {
-            gl::Viewport(0, 0, width as i32, height as i32);
+            gl::Viewport(0, 0, width, height);
         }
 
         Ok((
